core/src/main/java/hello/core/discount/RateDiscountPolicy.java
package hello.core.discount;
import hello.core.annotation.MainDiscountPolicy;
import hello.core.member.Grade;
import hello.core.member.Member;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.context.annotation.Primary;
import org.springframework.stereotype.Component;

@Component
@Primary // Rate가 무조건 먼저 선택되게 할 거야
@MainDiscountPolicy
public class RateDiscountPolicy implements DiscountPolicy {

    private int discountPercent = 10; //10% 할인

    @Override
    public int discount(Member member, int price) {
        if (member.getGrade() == Grade.VIP) { // 만약 VIP면 주문한 금액의 %를 할인
            return price * discountPercent / 100;
        } else {
            return 0;
        }
    }
}